+++ higlass/client.py
@@ -46,6 +46,7 @@ _datatype_default_track = {
     "gene-annotations": "horizontal-gene-annotations",
     "matrix": "heatmap",
     "vector": "horizontal-bar",
+    "multivec": "horizontal-multivec",
 }
 
 
@@ -74,7 +75,7 @@ class Track(Component):
     file_url: str
         An http accessible tileset file
     filetype : str
-        The type of the remote tilesets (e.g. 'bigwig' or 'cooler')
+        The type of the remote tilesets (e.g. 'bigwig', 'cooler', etc...)
     server : str, optional
         The server name (usually just 'localhost')
     height : int, optional
@@ -119,7 +120,7 @@ class Track(Component):
             self.conf["filetype"] = filetype
 
         if options is None:
-            options = {}
+            self.conf["options"] = {}
         else:
             self.conf["options"] = deepcopy(options)
 
@@ -177,6 +178,9 @@ class Track(Component):
 
         return CombinedTrack(new_tracks)
 
+    def __truediv__(self, other):
+        return DividedTrack(self, other,)
+
     @classmethod
     def from_dict(cls, conf):
         return cls(**conf)
@@ -186,6 +190,75 @@ class Track(Component):
 
     def copy(self):
         return Track(**self.to_dict())
+
+
+class DividedTrack(Track):
+    """A track representing one tileset divided by another.
+
+    Only works with some tileset types.
+    """
+
+    def __init__(
+        self, numerator, denominator, *args, **kwargs,
+    ):
+        """This track is created using two tilesets.
+
+        Parameters
+        ----------
+        numerator (tileset):
+            The tileset to be divided
+        denominator (tileset):
+            The tileset to divide by
+        """
+        if numerator.conf["type"] != denominator.conf["type"]:
+            raise ValueError(
+                f"Different track types: {numerator.conf['type']}, {denominator.conf['type']}"
+            )
+
+        if json.dumps(numerator.conf["options"]) != json.dumps(
+            denominator.conf["options"]
+        ):
+            logger.warn(
+                "Tracks have different options, so we're using the first track's"
+            )
+
+        numerator_server = numerator.conf["server"]
+        numerator_uuid = numerator.conf["tilesetUid"]
+
+        denominator_server = denominator.conf["server"]
+        denominator_uuid = denominator.conf["tilesetUid"]
+
+        track_type = numerator.conf["type"]
+        position = numerator.position
+        options = numerator.conf["options"]
+        height = numerator.conf["height"] if "height" in numerator.conf else None
+
+        data_config = {
+            "type": "divided",
+            "children": [
+                {"server": numerator_server, "tilesetUid": numerator_uuid},
+                {"server": denominator_server, "tilesetUid": denominator_uuid},
+            ],
+        }
+
+        super().__init__(
+            data=data_config,
+            type=track_type,
+            position=position,
+            options=options,
+            height=height,
+            *args,
+            **kwargs,
+        )
+
+    def change_attributes(self, **kwargs):
+        """
+        Change an attribute of this track and return a new copy.
+        """
+        conf = self.conf.copy()
+        conf.update(kwargs)
+
+        return Track(conf["type"]).from_dict(conf)
 
 
 class CombinedTrack(Track):
@@ -369,7 +442,13 @@ class View(Component):
                     klass = CombinedTrack
                 else:
                     klass = Track
-                self.add_track(track=klass.from_dict(track_conf), position=position)
+
+                # position has to be passed in as part of the parameter
+                # array so that the constructor can be called with it as
+                # a parameter
+                self.add_track(
+                    track=klass.from_dict({"position": position, **track_conf})
+                )
         return self
 
     def to_dict(self):